Moved common MCMC functions to their own file.
